+++ pkg/log/log.go
@@ -4,7 +4,12 @@ import (
 	"go.uber.org/zap"
 )
 
-var logger *zap.SugaredLogger
+type Logger struct {
+	*zap.SugaredLogger
+	eventLogger *zap.SugaredLogger
+}
+
+var logger *Logger
 
 func GetLogger() *zap.Logger {
 	return logger.Desugar()
@@ -14,6 +19,13 @@ func Init(opt *Options) {
 	if opt == nil {
 		return
 	}
-	logger = opt.buildLogger().Sugar()
+	logger = opt.buildLogger()
 }
 
+func (l *Logger) NewEvent() *Event {
+	return newEvent(l.eventLogger)
+}
+
+func NewEvent() *Event {
+	return logger.NewEvent()
+}